Added step transactions
Steps are added/subtracted from the bank on button presses. The bank is designed as follows: Each transaction, be it banking new steps on app launch or making a purchase, logs an entry into the database along with a timestamp of when the transaction occurred. The bank value, then, is the sum of all step entries in the database. Some form of "total steps" should be logged somewhere to ease the burden of summing all transactions each time an update is requested.